Add files via upload
The primary objective of performing Exploratory Data Analysis (EDA) on the Diwali sales data is to uncover meaningful insights and patterns that can inform business strategies and decision-making. By analyzing this dataset, we aim to understand the dynamics of consumer behavior, identify top-selling product categories, assess the impact of promotions.